Move some functions around
